att 1.3
Erros corrigidos, mas n sei se esta 100%

+++ arvoreBJava/arvoreB/Pagina.java
@@ -61,7 +61,7 @@ public class Pagina {
                 p.setChave(file.readInt(),i);
             }
             for(int i = 0;i<(2*ordem) +2;i++){
-                p.setOffset(file.readInt(),i);
+                p.setOffset(file.readLong(),i);
             }
         } catch (IOException e) {
             e.printStackTrace();
